+++ config/test.exs
@@ -2,8 +2,6 @@ import Config
 
 config :phoenix_test, :endpoint, PhoenixTest.WebApp.Endpoint
 
-config :logger, level: :warning
-
 config :phoenix_test, PhoenixTest.WebApp.Endpoint,
   server: true,
   http: [port: 4000],