+++ lib/phoenix_test/live.ex
@@ -42,15 +42,14 @@ defmodule PhoenixTest.Live do
   end
 
   def render_html(%{view: view, within: within}) do
-    case within do
-      :none ->
-        render(view)
+    html =
+      view
+      |> render()
+      |> Html.parse_fragment()
 
-      selector ->
-        view
-        |> render()
-        |> Query.find!(selector)
-        |> Html.raw()
+    case within do
+      :none -> html
+      selector when is_binary(selector) -> Html.all(html, selector)
     end
   end
 
@@ -70,17 +69,23 @@ defmodule PhoenixTest.Live do
       |> Query.find_by_role!(locator)
       |> Button.build(html)
 
-    click_button(session, button.selector, button.text)
+    handle_click_button(session, button)
   end
 
   def click_button(session, selector, text) do
-    html = render_html(session)
-    button = Button.find!(html, selector, text)
+    button =
+      session
+      |> render_html()
+      |> Button.find!(selector, text)
 
+    handle_click_button(session, button)
+  end
+
+  defp handle_click_button(session, button) do
     cond do
       Button.phx_click?(button) ->
         session.view
-        |> element(selector, text)
+        |> element(button.selector, button.text)
         |> render_click()
         |> maybe_redirect(session)
 
@@ -102,7 +107,8 @@ defmodule PhoenixTest.Live do
 
       true ->
         raise ArgumentError, """
-        Expected element with selector #{inspect(selector)} and text #{inspect(text)} to have a valid `phx-click` attribute or belong to a `form` element.
+        Expected element with selector #{inspect(button.selector)} and text
+          #{inspect(button.text)} to have a valid `phx-click` attribute or belong to a `form` element.
         """
     end
   end
@@ -189,6 +195,14 @@ defmodule PhoenixTest.Live do
     field = Field.find_checkbox!(html, input_selector, label, opts)
 
     cond do
+      # Support phx-click on checkboxes that have phx-value-key attributes too
+      Field.phx_click?(field) and Field.phx_value?(field) ->
+        session.view
+        |> element(field.selector)
+        |> render_click()
+        |> maybe_redirect(session)
+
+      # Support phx-click on checkboxes that aren't in forms
       Field.phx_click?(field) ->
         event = Html.attribute(field.parsed, "phx-click")
 
@@ -379,7 +393,8 @@ defmodule PhoenixTest.Live do
       |> render_html()
       |> Form.find!(selector)
 
-    form_data = select_form_data_to_submit(form, form_data)
+    form_data = remove_data_for_fields_that_have_been_removed(form_data, form)
+    form_data = FormData.merge(form.form_data, form_data)
 
     additional_data =
       if form.submit_button do
@@ -408,28 +423,12 @@ defmodule PhoenixTest.Live do
     end
   end
 
-  defp select_form_data_to_submit(form, form_data) do
-    element_names_present_in_final_form = Form.form_element_names(form)
-    form_data = remove_data_for_fields_that_have_been_removed(form_data, element_names_present_in_final_form)
+  defp remove_data_for_fields_that_have_been_removed(form_data, form) do
+    element_names = Form.form_element_names(form)
 
-    FormData.merge(form.form_data, form_data)
-  end
-
-  defp remove_data_for_fields_that_have_been_removed(form_data, element_names_present_in_final_form) do
     FormData.filter(form_data, fn %{name: name} ->
-      if index_based_field?(name) do
-        # We do not include index-based fields since those can be deleted and
-        # another (same name but diff value) take its place. Thus, we'll rely on
-        # the data found on the page.
-        false
-      else
-        name in element_names_present_in_final_form
-      end
+      name in element_names
     end)
-  end
-
-  defp index_based_field?(name) do
-    Regex.match?(~r/\[\d\]/, name)
   end
 
   def open_browser(%{view: view} = session, open_fun \\ &Phoenix.LiveViewTest.open_browser/1) do
@@ -490,7 +489,7 @@ defmodule PhoenixTest.Live do
         active_form? = form.selector == active_form.selector
         form_data = FormData.merge(form.form_data, if(active_form?, do: active_form.form_data, else: FormData.new()))
 
-        session.conn
+        %{session.conn | resp_body: html}
         |> PhoenixTest.Static.build()
         |> PhoenixTest.Static.submit_form(form.selector, form_data)