+++ lib/phoenix_test/static.ex
@@ -34,21 +34,20 @@ defmodule PhoenixTest.Static do
     |> render_html()
     |> Query.find("title")
     |> case do
-      {:found, element} -> Html.text(element)
+      {:found, element} -> Html.inner_text(element)
       _ -> nil
     end
   end
 
   def render_html(%{conn: conn, within: within}) do
-    case within do
-      :none ->
-        html_response(conn, conn.status)
+    html =
+      conn
+      |> html_response(conn.status)
+      |> Html.parse_document()
 
-      selector ->
-        conn
-        |> html_response(conn.status)
-        |> Query.find!(selector)
-        |> Html.raw()
+    case within do
+      :none -> html
+      selector when is_binary(selector) -> Html.all(html, selector)
     end
   end
 
@@ -235,9 +234,9 @@ defmodule PhoenixTest.Static do
 
     html =
       session.conn.resp_body
-      |> Floki.parse_document!()
-      |> Floki.traverse_and_update(&OpenBrowser.prefix_static_paths(&1, @endpoint))
-      |> Floki.raw_html()
+      |> Html.parse_document()
+      |> Html.postwalk(&OpenBrowser.prefix_static_paths(&1, @endpoint))
+      |> Html.raw()
 
     File.write!(path, html)